Run DirectoryChooser on FXApplicationThread -- fixes #339
During the launcher startup (splash screen) the DirectoryChooser dialogs are run offside the main application thread, therefore the dialog invocation was wrapped inside a task and scheduled via `Platform.runLater()`. However, this failed when called from the main application thread, as happened when trying to edit the directories from the settings menu.
The `DirectoryUtil` utility method now checks whether it is on the FX application thread or not, and properly decides how to run the directory chooser.