Added a `createPatch` task to the build script.
The new task tries to create a patch file between the last successful build artifact and the current build, based on the JOB_NAME. To do so, it will download the last successful artifact and create a delta patch file from the unzipped download and the unzipped distribution package (as created by `distZip` task). The generated patch file is then placed in the distribution directory.