Restructured Launcher Settings.
The goal down the road is to be able to have "temporary settings" in effect, e.g., start the launcher temporarily in offline mode or with a specific locale. The main problem currently is that the launcher settings are tied to the underlying properties file, i.e., changing the locale temporarily requires to change the property object.

Therefore, the main launcher settings which are stored in a Java properties file are separated from additional (temporary) settings. This is achieved by introducing a `LauncherSettingsDecorator` which allows to decorate the `BaseLauncherSettings` object with temporary settings.