Custom updater for Windows
When the launcher is installed with the NSIS-based installer (Windows) and it detects an update,
it will download and launch the new installer instead of trying to manually overwrite itself.
The update system has been refactored to allow multiple update methods; the launcher determines
how it should update itself by reading an updateStrategy file in its installation directory.
If this file isn't found, the default update strategy is used (self-update); otherwise, the
strategy to use is determined according to the file's content. For example, the NSIS installer
writes to this file to instruct the launcher to download and run the new installer when a new
version is found.
This system is designed to be extensible: different packaging methods can implement different
update strategies; for example, when the launcher is packaged for a package manager, the update
strategy could be to display a message asking the user to use the system package manager to get
the latest version.