Limit size of Warning button HBox
When the Warning was added, it was created with its own HBox to correctly place it to the left.
This HBox was overlaping all social buttons, making them unreachable.
As HBox with warning button had defined prefHeight to 64 but no maxWidth, this commit sets this property, efectivelly preventing the overlap.