Rename `DirectoryUtils` to `LauncherDirectoryUtils`
The utility methods for directories are very specific for the launcher
and even contain constants for the data directory name, cache directory,
or application directory name.
The new name should indicate this dedicated purpose.

`FileUtils` are more general and not tightly coupled to launcher
functionality. Some methods might be moved there in the future.