chore(actions): Build and upload assets on release
This action triggers when a release is published (Github event). The tagged commit is checked out and all assets are build using `gradlew allDistZip`.
This builds the zip bundles for all platforms we currently supports.

The resulting assets are uploaded to the release, where they can be downloaded by users.