Work in Progress: add slf4j-test
This is a slf4j provider made to write test assertions against.

But there can be only one slf4j provider on your classpath at the same time, otherwise who knows which will be used.

the remal.component-metadata plugin identifies a lot of these sorts of conflicts so they don't happen silently.

but that doesn't have spf4j-slf4j-test in its list of things it knows about, so we also need to specify that.

once you've identified conflicts, you have to resolve them, which takes ten lines for some reason, so I put it in a function.

debugging wasn't working very well on scripts included with `apply from`, and that's how I ended up writing a gradle plugin.