chore: Remove dead code (config package)
I always thought we were already using both Config and Settings side by
side, but that does not seem to be the case. Removing the whole package
still compile and runs the launcher fine.

I'll keep in mind that we had code here, and will probably salvage some
bits and pieces from it when refactoring the settings as we have them
right now.