feat: add Github repository adapter to fetch releases from Github
This will currently run into rate limit problems as we need to do a new request for each release to get the artifacts. I've proposed a PR to hub4j/github-api to fix that.
Another option would be to lazily fetch the assets/download URL, but that we way cannot guarantee that the desired asset is present in the release.