fix(repositories): make repository adapter more resilient
We were expecting that the Jenkins build would always have all required artifacts and information. We did not handle exceptional cases gracefully.

This change refactors the internals of JenkinsRepositoryAdapter to be more resilient by just skipping builds that don't match the requirements, e.g., that are missing artifact files.

Fixes #623