fix: append build number to version in JenkinsRepositoryAdapter
All preview builds of the Omega bundle triggered from engine's `develop`
branch have the same display version string (currently "alpha-18"). As
this version string is not unique, the launcher treats all those
releases as the same (at least in the UI).

This PR appends the build number to the version string in the
GameIdentifier to be able to distinguish the builds. It's only a quick
fix as the launcher should not mangle with version IDs on its own, but
just consume them from whatever repository source it connects to...

Quick fix for #630