feat: evolve `Settings` to an FX property-based settings object
This prepares `Settings` to take over the role of `LauncherSettings` as
source of truth for the launcher settings.

The information is stored in JavaFX properties wrapping the actual
value. The use of properties allows for reactive bindings to simplify
automatic updates of the UI.

When storig a (legacy) `LauncherSettings` object we also deserialize an
equivalent `Settings` object to JSON. Vice versa, we can also read
`Settings` from a JSON file.

The new `Settings` object is not yet used.