refactor: change API of Settings#load/store to take folder path instead of file
Since it is the decision of `Settings` what file types to write, the file path should not be passed in from the outside. Instead, just the folder path can be controlled.

As a consequence, the exact file name(s) are an implementation detail and no longer need to be exposed by `Settings`.