fix: handle unsuccessful requests in JenkinsClient better (#706)
A small change to handle unsuccessful requests via the JenkinsClient more gracefully.
Currently, a request failing with 4xx or 5xx will print a WARN log message and the whole stacktrace.
However, further up in the code we have the Response object and can check for the status code before continuing processing. With this change, we only log a single WARN message now containing the status code.