chore: restructure 'platform' package and add 'jre' package
Since we want to manage JREs for different platforms, the respective code needs to become a bit more elaborate. Therefore, introduce a new `platform` package with the `Platform` utility class and `OS` and `Arch` enums.

In addition, start working on `jre`-related classes to define and handle `JREArtefact`s to be downloaded and managed.