chore: make RepositoryManager a composite ReleaseRepository (#711)
The public interface of `RepositoryManager` is basically that of `ReleaseRepository`. Its purpose is to combine the different sources for game releases and hide where exactly a release is coming from.

Thus, make it implement `ReleaseRepository` and use the interface in more places instead of the specific class.

Also, drop the `...Adapter` suffixes as I'm not sure what they were intended to express anymore. The interface is called `ReleaseRepository` (with Adapter), and the two classes are implementing that interface.

In addition, this also deletes the unused class `JobResult`.